requirements.txt
langgraph>=0.2.0
litellm>=1.40.0
chromadb>=0.5.0
pyyaml>=6.0
pandas>=2.0
jupyter>=1.0
python-docx>=1.0
webvtt-py>=0.5
rich>=13.0
